refactor(DC-704): extract JWT authentication setup into AuthenticationSetup
